requirements.txt
spacy==3.7.2
numpy==1.26.4
scikit-learn==1.4.2
torch==2.2.1
datasets
joblib
transformers==4.38.2
accelerate==0.27.2
tabulate
wikipedia
python-dotenv
nltk
requests
fastapi
uvicorn
pydantic
pytest
httpx
sympy
python-multipart
supabase